+++ org.qiweb/org.qiweb.test/src/main/java/org/qiweb/test/QiWebHttpTest.java
@@ -20,12 +20,15 @@ import org.junit.Before;
 import org.qiweb.api.Mode;
 import org.qiweb.runtime.ApplicationInstance;
 import org.qiweb.runtime.ConfigInstance;
+import org.qiweb.runtime.ConfigKeys;
+import org.qiweb.runtime.CryptoInstance;
 import org.qiweb.runtime.routes.RoutesConfProvider;
 import org.qiweb.runtime.routes.RoutesProvider;
 import org.qiweb.server.netty.NettyServer;
 import org.qiweb.spi.ApplicationSPI;
 import org.qiweb.spi.server.HttpServer;
 
+import static java.util.Collections.singletonMap;
 import static org.qiweb.runtime.ConfigKeys.QIWEB_HTTP_ADDRESS;
 import static org.qiweb.runtime.ConfigKeys.QIWEB_HTTP_PORT;
 import static org.qiweb.test.QiWebTestHelper.setupRestAssuredDefaults;
@@ -81,6 +84,16 @@ public class QiWebHttpTest
                       ? configurationResourceName()
                       : configurationResourceNameOverride;
         ConfigInstance config = new ConfigInstance( classLoader, conf );
+        try
+        {
+            config.string( ConfigKeys.APP_SECRET );
+        }
+        catch( com.typesafe.config.ConfigException.Missing noAppSecret )
+        {
+            System.out.println( "Application has no 'app.secret', generating a random one for test mode!" );
+            String secret = CryptoInstance.newRandomSecret256BitsHex();
+            config = new ConfigInstance( classLoader, conf, null, null, singletonMap( "app.secret", secret ) );
+        }
         RoutesProvider routesProvider = routesProviderOverride == null
                                         ? routesProvider()
                                         : routesProviderOverride;
@@ -138,7 +151,7 @@ public class QiWebHttpTest
      */
     protected String configurationResourceName()
     {
-        return "application.conf";
+        return null;
     }
 
     /**